pyproject.toml
[project]
name = "ovi-voice-assistant"
version = "0.1.0"
description = "Ovi — AI-powered voice assistant with streaming STT, TTS"
readme = "README.md"
authors = [
    { name = "Bryan Furia", email = "bryan.furia@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "faster-whisper==1.2.1",
    "piper-tts==1.4.2",
    "openai-agents==0.13.4",
    "aioesphomeapi==44.6.2",
    "numpy==2.4.4",
    "audresample==1.3.5",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "lc3py>=1.1.0",
    "opuslib>=3.0.1",
    "bleak>=0.22.0",
    "fastembed>=0.6.0",
    "pocket-tts==1.1.1",
    "onnxruntime>=1.18.0",
    "ytmusicapi>=1.8.0",
    "yt-dlp>=2025.1.0",
    "playwright>=1.49.0",
    "websockets>=14.0",
]

[dependency-groups]
dev = ["esphome==2026.3.2", "pytest>=8.0", "pytest-asyncio>=0.25", "ruff>=0.11"]

[project.scripts]
ovi = "ovi_voice_assistant.__main__:main"

[tool.pytest.ini_options]
testpaths = ["src"]
python_files = ["*_test.py"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py314"
line-length = 88

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific rules
]
ignore = [
    "E501",   # line length handled by ruff format
    "SIM105", # contextlib.suppress isn't always clearer than try/except/pass
    "SIM108", # ternary isn't always clearer, especially with comments
]

[build-system]
requires = ["uv_build>=0.9.26"]
build-backend = "uv_build"
